cfo_agent/requirements.txt
flask==2.3.3
gunicorn==21.2.0
google-cloud-secretmanager==2.16.0
google-cloud-storage==2.9.0
pandas==2.0.3
numpy==1.24.4
plotly==5.15.0
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
typing-extensions==4.7.1
mypy==1.5.1
mypy-extensions==1.0.0
pydantic==2.1.1
pydantic-core==2.4.0
pytest==7.4.0
pytest-mock==3.12.0 